target/classes/static/interact_with_api/requirements.txt
fastapi
uvicorn
gunicorn
openai
python-dotenv
pydantic